requirements.txt
pyglet==1.5.0
bresenham==0.2.1
pypng==0.0.20
numpy==1.18.1
